requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
Pillow>=10.0.0
tenacity>=8.2.3
tqdm>=4.66.0
numpy>=2.0.0
